pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quant-lab"
version = "0.1.0"
description = "AI Quant Research Lab - Institution-grade quantitative finance research infrastructure"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "numpy>=1.26.0,<2.0",
    "pandas>=2.1.0",
    "scipy>=1.11.0",
    "python-dateutil>=2.8.2",
    "yfinance>=0.2.31",
    "pyarrow>=14.0.0",
    "torch>=2.2.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.2.0",
    "mlflow>=2.10.0",
    "plotly>=5.18.0",
    "jinja2>=3.1.0",
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "hmmlearn>=0.3.0",
    "tqdm>=4.66.0",
    "structlog>=23.2.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["quant_lab"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks integration tests",
    "gpu: marks tests requiring GPU",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
